Parser (#118)
* Move parser to its own module

* Add ORIGIN to env vars

---------
