+++ backend/program/scrapers/__init__.py
@@ -1,11 +1,11 @@
 from datetime import datetime
-from pydantic import BaseModel, Extra
+from pydantic import BaseModel
 from utils.service_manager import ServiceManager
+from utils.settings import settings_manager as settings
 from utils.logger import logger
 from .torrentio import Torrentio
 from .orionoid import Orionoid
 from .jackett import Jackett
-from utils.settings import settings_manager as settings
 
 
 class ScrapingConfig(BaseModel):
