+++ entrypoint.sh
@@ -15,7 +15,6 @@ fi
 : ${USERNAME:=iceberg}
 : ${GROUPNAME:=iceberg}
 
-
 if ! getent group ${PGID} >/dev/null; then
     addgroup -g $PGID $GROUPNAME > /dev/null
 else
@@ -30,5 +29,7 @@ fi
 
 chown -R ${USERNAME}:${GROUPNAME} /iceberg
 
+ORIGIN=${ORIGIN:-http://localhost:3000}
+
 echo "Container Initialization complete."
-exec su -m $USERNAME -c 'cd backend && source /venv/bin/activate && exec python /iceberg/backend/main.py & ORIGIN=http://localhost:3000 node /iceberg/frontend/build'
+exec su -m $USERNAME -c 'cd backend && source /venv/bin/activate && exec python /iceberg/backend/main.py & ORIGIN=$ORIGIN node /iceberg/frontend/build'
